forex_bot/requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
arch>=6.0.0
yfinance>=0.2.0
optuna>=3.0.0
